BannerColor rgb param type fix byte -> int

+++ Assignment2/src/academy/pocu/comp2500/assignment2/BannerColor.java
@@ -3,7 +3,7 @@ package academy.pocu.comp2500.assignment2;
 public class BannerColor {
     protected int rgba;
 
-    public BannerColor(byte r, byte g, byte b) {
+    public BannerColor(int r, int g, int b) {
         rgba = (r << 24 | g << 16 | b << 8);
     }
 }
